feat(dial): exposing connection handle flow
